texturas

+++ app/src/main/java/util/Time.java
@@ -2,7 +2,7 @@ package util;
 
 public class Time {
     private static double startTime = System.nanoTime();
-    public static double getTime() {
-        return (System.nanoTime() - startTime) / 1e9;
+    public static float getTime() {
+        return (float)((System.nanoTime() - startTime) / 1e9);
     }
 }
